Add timeout functionality when matching connections
An issue in the cockroachdb repo (https://github.com/cockroachdb/cockroach/issues/143785)
uncovered an undesired behaviour in the goroutines handled by cmux.

Particularly, cmux can hang forever on trying to match a connection.

This can be resolved by users of cmux as illustrated in
https://github.com/cockroachdb/cockroach/pull/144170. However, it is an
unorthodox solution.

This commit adds functionality to cmux to optionally specify a timeout on
matching (reading) connections.